+++ Consul/Interfaces/IEventEndpoint.cs
@@ -23,12 +23,11 @@ namespace Consul
 {
     public interface IEventEndpoint
     {
-        Task<WriteResult<string>> Fire(UserEvent ue);
-        Task<WriteResult<string>> Fire(UserEvent ue, WriteOptions q);
+        Task<WriteResult<string>> Fire(UserEvent ue, CancellationToken ct = default(CancellationToken));
+        Task<WriteResult<string>> Fire(UserEvent ue, WriteOptions q, CancellationToken ct = default(CancellationToken));
         ulong IDToIndex(string uuid);
-        Task<QueryResult<UserEvent[]>> List();
-        Task<QueryResult<UserEvent[]>> List(string name);
-        Task<QueryResult<UserEvent[]>> List(string name, QueryOptions q);
-        Task<QueryResult<UserEvent[]>> List(string name, QueryOptions q, CancellationToken ct);
+        Task<QueryResult<UserEvent[]>> List(CancellationToken ct = default(CancellationToken));
+        Task<QueryResult<UserEvent[]>> List(string name, CancellationToken ct = default(CancellationToken));
+        Task<QueryResult<UserEvent[]>> List(string name, QueryOptions q, CancellationToken ct = default(CancellationToken));
     }
 }