+++ Consul/Session.cs
@@ -69,7 +69,7 @@ namespace Consul
             return objectType == typeof(SessionBehavior);
         }
     }
-    
+
     public class SessionExpiredException : Exception
     {
         public SessionExpiredException() { }
@@ -228,18 +228,18 @@ namespace Consul
         /// <param name="se">The SessionEntry options to use</param>
         /// <returns>A write result containing the new session ID</returns>
 
-        public Task<WriteResult<string>> Create()
+        public Task<WriteResult<string>> Create(CancellationToken ct = default(CancellationToken))
         {
-            return Create(null, WriteOptions.Default);
+            return Create(null, WriteOptions.Default, ct);
         }
 
         /// <summary>
         /// Create makes a new session with default options.
         /// </summary>
         /// <returns>A write result containing the new session ID</returns>
-        public Task<WriteResult<string>> Create(SessionEntry se)
+        public Task<WriteResult<string>> Create(SessionEntry se, CancellationToken ct = default(CancellationToken))
         {
-            return Create(se, WriteOptions.Default);
+            return Create(se, WriteOptions.Default, ct);
         }
 
         /// <summary>
@@ -248,17 +248,17 @@ namespace Consul
         /// <param name="se">The SessionEntry options to use</param>
         /// <param name="q">Customized write options</param>
         /// <returns>A write result containing the new session ID</returns>
-        public async Task<WriteResult<string>> Create(SessionEntry se, WriteOptions q)
+        public async Task<WriteResult<string>> Create(SessionEntry se, WriteOptions q, CancellationToken ct = default(CancellationToken))
         {
-            var res = await _client.Put<SessionEntry, SessionCreationResult>("/v1/session/create", se, q).Execute().ConfigureAwait(false);
+            var res = await _client.Put<SessionEntry, SessionCreationResult>("/v1/session/create", se, q).Execute(ct).ConfigureAwait(false);
             return new WriteResult<string>(res, res.Response.ID);
         }
         /// <summary>
         /// CreateNoChecks is like Create but is used specifically to create a session with no associated health checks.
         /// </summary>
-        public Task<WriteResult<string>> CreateNoChecks()
+        public Task<WriteResult<string>> CreateNoChecks(CancellationToken ct = default(CancellationToken))
         {
-            return CreateNoChecks(null, WriteOptions.Default);
+            return CreateNoChecks(null, WriteOptions.Default, ct);
         }
 
         /// <summary>
@@ -266,9 +266,9 @@ namespace Consul
         /// </summary>
         /// <param name="se">The SessionEntry options to use</param>
         /// <returns>A write result containing the new session ID</returns>
-        public Task<WriteResult<string>> CreateNoChecks(SessionEntry se)
+        public Task<WriteResult<string>> CreateNoChecks(SessionEntry se, CancellationToken ct = default(CancellationToken))
         {
-            return CreateNoChecks(se, WriteOptions.Default);
+            return CreateNoChecks(se, WriteOptions.Default, ct);
         }
 
         /// <summary>
@@ -277,7 +277,7 @@ namespace Consul
         /// <param name="se">The SessionEntry options to use</param>
         /// <param name="q">Customized write options</param>
         /// <returns>A write result containing the new session ID</returns>
-        public Task<WriteResult<string>> CreateNoChecks(SessionEntry se, WriteOptions q)
+        public Task<WriteResult<string>> CreateNoChecks(SessionEntry se, WriteOptions q, CancellationToken ct = default(CancellationToken))
         {
             if (se == null)
             {
@@ -292,7 +292,7 @@ namespace Consul
                 Node = se.Node,
                 TTL = se.TTL
             };
-            return Create(noChecksEntry, q);
+            return Create(noChecksEntry, q, ct);
         }
 
         /// <summary>
@@ -300,9 +300,9 @@ namespace Consul
         /// </summary>
         /// <param name="id">The session ID to destroy</param>
         /// <returns>A write result containing the result of the session destruction</returns>
-        public Task<WriteResult<bool>> Destroy(string id)
+        public Task<WriteResult<bool>> Destroy(string id, CancellationToken ct = default(CancellationToken))
         {
-            return Destroy(id, WriteOptions.Default);
+            return Destroy(id, WriteOptions.Default, ct);
         }
 
         /// <summary>
@@ -311,9 +311,9 @@ namespace Consul
         /// <param name="id">The session ID to destroy</param>
         /// <param name="q">Customized write options</param>
         /// <returns>A write result containing the result of the session destruction</returns>
-        public Task<WriteResult<bool>> Destroy(string id, WriteOptions q)
+        public Task<WriteResult<bool>> Destroy(string id, WriteOptions q, CancellationToken ct = default(CancellationToken))
         {
-            return _client.Put<object, bool>(string.Format("/v1/session/destroy/{0}", id), q).Execute();
+            return _client.Put<object, bool>(string.Format("/v1/session/destroy/{0}", id), q).Execute(ct);
         }
 
         /// <summary>
@@ -321,9 +321,9 @@ namespace Consul
         /// </summary>
         /// <param name="id">The session ID to look up</param>
         /// <returns>A query result containing the session information, or an empty query result if the session entry does not exist</returns>
-        public Task<QueryResult<SessionEntry>> Info(string id)
+        public Task<QueryResult<SessionEntry>> Info(string id, CancellationToken ct = default(CancellationToken))
         {
-            return Info(id, QueryOptions.Default);
+            return Info(id, QueryOptions.Default, ct);
         }
 
         /// <summary>
@@ -332,9 +332,9 @@ namespace Consul
         /// <param name="id">The session ID to look up</param>
         /// <param name="q">Customized query options</param>
         /// <returns>A query result containing the session information, or an empty query result if the session entry does not exist</returns>
-        public async Task<QueryResult<SessionEntry>> Info(string id, QueryOptions q)
+        public async Task<QueryResult<SessionEntry>> Info(string id, QueryOptions q, CancellationToken ct = default(CancellationToken))
         {
-            var res = await _client.Get<SessionEntry[]>(string.Format("/v1/session/info/{0}", id), q).Execute().ConfigureAwait(false);
+            var res = await _client.Get<SessionEntry[]>(string.Format("/v1/session/info/{0}", id), q).Execute(ct).ConfigureAwait(false);
             return new QueryResult<SessionEntry>(res, res.Response != null && res.Response.Length > 0 ? res.Response[0] : null);
         }
 
@@ -342,9 +342,9 @@ namespace Consul
         /// List gets all active sessions
         /// </summary>
         /// <returns>A query result containing list of all sessions, or an empty query result if no sessions exist</returns>
-        public Task<QueryResult<SessionEntry[]>> List()
+        public Task<QueryResult<SessionEntry[]>> List(CancellationToken ct = default(CancellationToken))
         {
-            return List(QueryOptions.Default);
+            return List(QueryOptions.Default, ct);
         }
 
         /// <summary>
@@ -352,9 +352,9 @@ namespace Consul
         /// </summary>
         /// <param name="q">Customized query options</param>
         /// <returns>A query result containing the list of sessions, or an empty query result if no sessions exist</returns>
-        public Task<QueryResult<SessionEntry[]>> List(QueryOptions q)
+        public Task<QueryResult<SessionEntry[]>> List(QueryOptions q, CancellationToken ct = default(CancellationToken))
         {
-            return _client.Get<SessionEntry[]>("/v1/session/list", q).Execute();
+            return _client.Get<SessionEntry[]>("/v1/session/list", q).Execute(ct);
         }
 
         /// <summary>
@@ -362,9 +362,9 @@ namespace Consul
         /// </summary>
         /// <param name="node">The node ID</param>
         /// <returns>A query result containing the list of sessions, or an empty query result if no sessions exist</returns>
-        public Task<QueryResult<SessionEntry[]>> Node(string node)
+        public Task<QueryResult<SessionEntry[]>> Node(string node, CancellationToken ct = default(CancellationToken))
         {
-            return Node(node, QueryOptions.Default);
+            return Node(node, QueryOptions.Default, ct);
         }
 
         /// <summary>
@@ -373,9 +373,9 @@ namespace Consul
         /// <param name="node">The node ID</param>
         /// <param name="q">Customized query options</param>
         /// <returns>A query result containing the list of sessions, or an empty query result if no sessions exist</returns>
-        public Task<QueryResult<SessionEntry[]>> Node(string node, QueryOptions q)
+        public Task<QueryResult<SessionEntry[]>> Node(string node, QueryOptions q, CancellationToken ct = default(CancellationToken))
         {
-            return _client.Get<SessionEntry[]>(string.Format("/v1/session/node/{0}", node), q).Execute();
+            return _client.Get<SessionEntry[]>(string.Format("/v1/session/node/{0}", node), q).Execute(ct);
         }
 
         /// <summary>
@@ -383,9 +383,9 @@ namespace Consul
         /// </summary>
         /// <param name="id">The session ID to renew</param>
         /// <returns>An updated session entry</returns>
-        public Task<WriteResult<SessionEntry>> Renew(string id)
+        public Task<WriteResult<SessionEntry>> Renew(string id, CancellationToken ct = default(CancellationToken))
         {
-            return Renew(id, WriteOptions.Default);
+            return Renew(id, WriteOptions.Default, ct);
         }
 
         /// <summary>
@@ -394,9 +394,9 @@ namespace Consul
         /// <param name="id">The session ID to renew</param>
         /// <param name="q">Customized write options</param>
         /// <returns>An updated session entry</returns>
-        public async Task<WriteResult<SessionEntry>> Renew(string id, WriteOptions q)
+        public async Task<WriteResult<SessionEntry>> Renew(string id, WriteOptions q, CancellationToken ct = default(CancellationToken))
         {
-            var res = await _client.Put<object, SessionEntry[]>(string.Format("/v1/session/renew/{0}", id), q).Execute().ConfigureAwait(false);
+            var res = await _client.Put<object, SessionEntry[]>(string.Format("/v1/session/renew/{0}", id), q).Execute(ct).ConfigureAwait(false);
             if (res.StatusCode == HttpStatusCode.NotFound)
             {
                 throw new SessionExpiredException(string.Format("Session expired: {0}", id));